feat: correct type of Integer to Long

+++ src/main/java/com/luan/desafio/desafioestagio/repository/ClienteRepository.java
@@ -3,5 +3,5 @@ package com.luan.desafio.desafioestagio.repository;
 import com.luan.desafio.desafioestagio.model.Cliente;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+public interface ClienteRepository extends JpaRepository<Cliente, Long> {
 }
